Comments update

+++ grisera_api/mongo_service/mongo_api_service.py
@@ -2,6 +2,7 @@ from typing import Union
 from pydantic import BaseModel
 from pymongo import MongoClient
 from bson import ObjectId
+from grisera_api.mongo_service.utils import mongo_deep_iteration
 
 from models.not_found_model import NotFoundByIdModel
 from mongo_service.collection_mapping import get_collection_name
@@ -10,7 +11,7 @@ from mongo_service.mongodb_api_config import mongo_api_address, mongo_database_n
 
 class MongoApiService:
     """
-    Object that handles communication with mongodb
+    Object that handles direct communication with mongodb
     """
 
     MONGO_ID_FIELD = "_id"
@@ -25,17 +26,17 @@ class MongoApiService:
 
     def create_document(self, data_in: BaseModel):
         """
-        Create new document
+        Create new document. Id fields are converted to ObjectId type.
         """
         collection_name = get_collection_name(type(data_in))
         data_as_dict = data_in.dict()
-        self._fix_query_ids(data_as_dict)
+        self._fix_input_ids(data_as_dict)
         created_id = self.db[collection_name].insert_one(data_as_dict).inserted_id
         return str(created_id)
 
     def get_document(self, id: Union[str, int], model_class, *args, **kwargs):
         """
-        Load single document
+        Load single document. Output id fields are converted from ObjectId type to str.
         """
         collection_name = get_collection_name(model_class)
         result_dict = self.db[collection_name].find_one(
@@ -53,10 +54,10 @@ class MongoApiService:
 
     def get_documents(self, model_class, query: dict = {}, *args, **kwargs):
         """
-        Load many documents
+        Load many documents. Output id fields are converted from ObjectId type to str.
         """
         collection_name = get_collection_name(model_class)
-        self._fix_query_ids(query)
+        self._fix_input_ids(query)
         results = list(self.db[collection_name].find(query, *args, **kwargs))
 
         [self._update_mongo_output_id(result) for result in results]
@@ -65,7 +66,7 @@ class MongoApiService:
 
     def update_document(self, id: Union[str, int], data_to_update: BaseModel):
         """
-        Update document
+        Update document.
         """
         collection_name = get_collection_name(type(data_to_update))
         data_as_dict = data_to_update.dict()
@@ -75,7 +76,7 @@ class MongoApiService:
         self, collection_name: str, id: Union[str, int], new_document: dict
     ):
         """
-        Update document with new document as dict
+        Update document with new document as dict. Id fields are converted to ObjectId type.
         """
         self._update_mongo_input_id(new_document)
         id = ObjectId(id)
@@ -86,64 +87,65 @@ class MongoApiService:
 
     def delete_document(self, object_to_delete: BaseModel):
         """
-        Delete document in collection
+        Delete document in collection. Given model must have id field.
         """
-        id = ObjectId(object_to_delete.id)
+        id_str = getattr(object_to_delete, self.MODEL_ID_FIELD, None)
+        if id_str is None:
+            raise TypeError(
+                f"Given model object does not have '{self.MODEL_ID_FIELD}' field"
+            )
+        id = ObjectId(id_str)
         collection_name = get_collection_name(type(object_to_delete))
         self.db[collection_name].delete_one({self.MONGO_ID_FIELD: id})
         return id
 
-    @staticmethod
-    def _add_additional_properties_to_dict(data_dict: dict):
-        if data_dict["additional_properties"] is None:
-            return
-        for additional_property in data_dict["additional_properties"]:
-            data_dict[additional_property["key"]] = data_dict[
-                additional_property["value"]
-            ]
-        del data_dict["additional_properties"]
-
     def _update_mongo_input_id(self, mongo_input: dict):
+        """
+        Mongo documents id fields are '_id' while models fields are 'id'. Here id field is
+        renamed and other id fields (relation fields) types are converted.
+        """
         if self.MODEL_ID_FIELD in mongo_input:
             mongo_input[self.MONGO_ID_FIELD] = ObjectId(
                 mongo_input[self.MODEL_ID_FIELD]
             )
         del mongo_input[self.MODEL_ID_FIELD]
-        self._fix_query_ids(mongo_input)
+        self._fix_input_ids(mongo_input)
 
     def _update_mongo_output_id(self, mongo_output: dict):
+        """
+        Mongo documents id fields are '_id' while models fields are 'id'. Here id field is
+        renamed and other id fields (relation fields) types are converted.
+        """
         if self.MONGO_ID_FIELD in mongo_output:
             mongo_output[self.MODEL_ID_FIELD] = str(mongo_output[self.MONGO_ID_FIELD])
         del mongo_output[self.MONGO_ID_FIELD]
         self._fix_output_ids(mongo_output)
 
-    def _fix_query_ids(self, query):
-        for field, value in query.items():
-            if type(value) is dict:
-                self._fix_query_ids(value)
-            elif self._field_is_id(field):
-                query[field] = ObjectId(query[field])
+    @mongo_deep_iteration
+    def _fix_input_ids(self, field):
+        """
+        Mongo uses ObjectId in id fields, while models use int/str. This function
+        performs conversion on each id field in input query.
+        """
+        if self._field_is_id(field):
+            return ObjectId(field)
+        return field
 
-    def _fix_output_ids(self, result):
+    @mongo_deep_iteration
+    def _fix_output_ids(self, field):
         """
-        Perform deep iteration over query result and parser all ObjectId's fields to str
+        Mongo uses ObjectId in id fields, while models use int/str. This function
+        performs conversion on each id field in output dict.
         """
-        if type(result) is not dict:
-            return
-        for field, value in result.items():
-            if type(value) is dict:
-                self._fix_output_ids(value)
-            elif type(value) is list:
-                for list_elem in value:
-                    self._fix_output_ids(list_elem)
-            elif self._field_is_id(field):
-                result[field] = str(result[field])
+        if self._field_is_id(field):
+            return str(field)
+        return field
 
     @staticmethod
     def _field_is_id(field):
         if type(field) is not str:
             return False
-        return field == "id" or (len(field) >= 3 and field[-3:] in ("_id", ".id"))
+        return field == "id" or field[-3:] in ("_id", ".id")
 
 
 mongo_api_service = MongoApiService()
